fix(MTRNIX-352): move BootstrapStateStore to workspaces/bootstrap/store.py (L1->L3 cleanup)
