+++ src/metatron/workspaces/bootstrap/runner.py
@@ -19,7 +19,7 @@ if TYPE_CHECKING:
 
     from metatron.core.config import Settings
     from metatron.core.interfaces import ConnectorInterface
-    from metatron.storage.bootstrap_state import BootstrapStateStore
+    from metatron.workspaces.bootstrap.store import BootstrapStateStore
 
 logger = structlog.get_logger(__name__)
