+++ tests/integration/workspaces/test_delete_cascade.py
@@ -17,7 +17,7 @@ from sqlalchemy.ext.asyncio import create_async_engine
 
 from metatron.chat.persistence import ChatPersistence
 from metatron.core.config import get_settings
-from metatron.storage.bootstrap_state import BootstrapStateStore
+from metatron.workspaces.bootstrap.store import BootstrapStateStore
 from metatron.workspaces.manager import WorkspaceManager
 
 pytestmark = pytest.mark.integration
